start splitting apart into multiple files (command and separate lib)
